HTML
This commit restructures the repository so that it can (hopefully) be
used as a static site. For the moment the goal is publish the static
site in site. Each of the supplied HTML files had its filename
normalized and it was dropped into the site/plans directory. The HTML
files were cleaned with BeautifulSoup and saved with a similar filename
for now, instead of overwriting the original.